LABSPLAT-14 // LABSPLAT-2 An interesting step.  This is sort of a first pass at something that might be useful, but it's essentially a prototype.
It will build a VPC and spin up N management hosts. The next step is to get Ansible bootstrapped on a management host or come up with a nice way of jumping through the management host to run play books inside the VPC for any EC2 instance on a private subnet.

There is no NAT instance yet, instances in public subnets get access out through a internet gateway.

A new default security group is created in addition to an SSH one. The special replacement default SG is needed because of the way Cloudformation works.